Fix fetch* to work as specified in PEP249
I had misunderstood how the three fetch methods were supposed to interact, I thought each one would start a new iteration from the first row, but that's not the case. The hint is in the class name: cursor. The cursor is a cursor over the result set, and calling fetchone fetches one row, fetchmany fetches many, and fetchall fetches the remaining rows. The three methods can be called in any combination and sequence.

To implement this I moved the fetching to a separate class that implements Iterator. The cursor creates an instance as soon as the query has reached the SUCCEEDED state, and then the instance lazily loads pages as needed while it's being iterated. It also has two methods to get column metadata and the update count, and these just load the first page of results.